coloring

+++ Row.cpp
@@ -12,7 +12,7 @@ Row::Row(Grid& grid, size_t index)
     if (!_cells.empty())
     {
         _rect = (*_cells.begin()).rect();
-        _rect.setWidth((int)(_cells.size() * grid.columnWidth()));
+        _rect.setWidth((qreal)(_cells.size() * grid.columnWidth()));
         //qDebug("Row %zu (%d,%d) (%d,%d)", index, _rect.left(), _rect.top(), _rect.right(), _rect.bottom());
     }
 }
